+++ src/addons/display_commander/latent_sync/latent_sync_manager.cpp
@@ -1,7 +1,0 @@
-#include "latent_sync_manager.hpp"
-
-namespace dxgi::latent_sync {
-
-LatentSyncManager::LatentSyncManager() {}
-
-}  // namespace dxgi::latent_sync